Changed view matrix calculation to fix related problems.
Changed the view matrix to be the inverse of the Camera View matrix in the FREE_LOOK camera.
Updated the LOOK_AT view matrix to match an equivalent FREE_LOOK camera.
Updated sorting of renderables to work with the new view space.
Updated shader viewport uniform translation to match new view space.

Change-Id: Ideeacd1990ee4ebde715362e8006b0f2adebda91
Signed-off-by: Francisco Santos <f1.santos@samsung.com>