pan smoothing and prediction
Rebase of this change:

This patch keeps track of a history of the the last 60 touches.
The time stamp of the update (which is a pretty regular 16.7ms),
queries the history for the touch point at the update timestamp
minus a constant (e.g. 16.7ms). The history returns a touch
point which is the linearly interpolated value between the two
nearest touches in the sample. Thus reflecting where the touch
would've been at fixed intervals of a video frame.

-- Reverted to just blending the last two touch video-frames
Appears that most of the problem is noise in the touch positions
rather than their alignment with the frame interval.

Signed-off-by: Paul Wisbey <p.wisbey@samsung.com>