(Image) Add support for 9-patch images
[Issue#]       N/A
[Problem]      There is no support for 9 patch images with the
stretch / child info information stored in a 1 pixel border within
the image
[Cause]        N/A
[Solution]     Added new class NinePatchImage, which is automatically
used by Image::New when a 9 patch filename is seen.

Adding a 9 patch image to an ImageActor via an Image handle will generate
an alternative BitmapImage with the border cropped out. It will set up the
ImageActor to use the 9 patch style and set the stretch areas up appropriately.

Adding a 9 patch image to an ImageActor via a NinePatchImage handle will
use the image as is (i.e. without the border cropped out).

[Verification] Build Repo

Change-Id: I8fa5cc34220ea9abcf085c600279ef92bc12da00
Signed-off-by: David Steele <david.steele@partner.samsung.com>