(HitTest) Ensure non-renderable actors are hittable even if outside the stencil
We do not expect to receive a hit outside the stencil area for a rendered actor. However, for
non-renderable actors, we cannot see these anyway and so these should still be hittable even if we
haven't hit the stencil area. This would allow controls like scroll-view to work with stencils
where we only want the stencil to affect the children of scroll-view.

Change-Id: Ic8e9d00a4125c94972ad2cfe298fae65313ad7dd
Signed-off-by: Adeel Kazmi <adeel.kazmi@samsung.com>