Added connection for sampler texture
Adding an object connection for the sampler texture enables the
material to be informed when the sampler texture changes.

This is chained up to the renderable attachment, which re-sends
the data providers to the renderer, updating the texture.

Note, doubleBufferedProperty[i] = value does not enable the aging
behaviour, whereas doubleBufferedProperty.Set(i, value) does.

Change-Id: I3c43839e86cbc59a6c47598d287bfae512b18639
Signed-off-by: David Steele <david.steele@partner.samsung.com>