Ensuring GL Sync is only used for RefreshOnce render tasks
GL Sync is currently used when a framebuffer is backed with a native
image.  It is tracked using a RenderTracker, which is created in
SceneGraph::RenderTask::PrepareRenderInstruction.  In
RenderManager::DoRender, if the render tracker is available, then this
creates a sync object.  Consequently, if the render task is a
RefreshAlways task, then a sync object is created each frame.

This change ensures that a render tracker is only created for a
RefreshOnce render tracker, and that the sync object is created only
once, rather than every frame.

Change-Id: I93f5c2649c58cc3100b4ddc8546cf023cd728e5a
Signed-off-by: David Steele <david.steele@samsung.com>