Added integer keys for custom properties.
Custom properties (those that are registered, rather than predefined
in the DEFAULT range), can now also be registered with an integer key
in addition to the name.

The key can then be used to retrieve the property index. This is of
use for renderer properties that are well specified in toolkit.

Change-Id: Ieda47070f30f5cb53514c9358843f24868a3da3e
Signed-off-by: David Steele <david.steele@samsung.com>