Enabling animation of Renderer properties when offstage
If the renderer exists, but hasn't been staged yet, then the scene
graph object also exists, but cannot be accessed because of a spurious
check for OnStage().

Removing the OnStage check to allow animations to run on Renderers
that haven't 'quite' been staged but are very likely to in the near
future.

Change-Id: I219cca9e08ebb6c7c5b2b8e7f33fd8bbb24b47ae
Signed-off-by: David Steele <david.steele@samsung.com>