Added travis.yml for auto-building patches
The Travis build runs in an Ubuntu 14.04 VM, and has the necessary dependencies
to build dali-core and to build and execute the test cases (this does not require
testkit-lite anymore, but does still use it's scripts)

After building and running, it concatenates the coverage .info files
and uploads them to coveralls.io/github/dalihub/dali-core.

Updated coverage script to stop running genhtml (to reduce
time when uploading lcov data to coveralls).

Change-Id: I59bb8a4c043ffd84141b897070121d63f3132bed
Signed-off-by: David Steele <david.steele@samsung.com>