Fix memory leak, scene graph layers are never deleted from memory
Reported by Valgrind.
We have a memory pool for Nodes which works fine
for creating /deleting nodes.

Scene Graph Layers inherit from Nodes, but don't use
the memory pool.

However when either a Node or a Layer is deleted, the overloaded
Node operator delete( void* ptr ) is called, which tries to free
the Node from the memory pool.
Unfortunately for Layer it was never in the memory pool so no
memory is free'd.

Fix includes:
node destructor is protected
layer destructor is private
node delete operator removed

Change-Id: Icfe83f35b4f29d5b774cde392ff976ad299481c5