Fix Message Buffer leak and added DEBUG_ASSERT for leaking nodes
Current message queue is not cleared when update is destroyed
This is 1 cause of Node leak.
Second cause of Node leak is InstallRootMessage passes a layer
as a raw pointer. If update doesn't run ( doesn't run in the test
cases), then we leak the layer (node).

Also added DEBUG_ASSERT to check for leaking scene graph Nodes

Change-Id: Ibfbb73ee840cbeaf33104da6769945cf9b3aafbc