Added new mechanism to order actor depths
At the end of each frame, if the actor tree was changed, or the
sibling order of any actor was modified, then the actor tree is
traversed to determine the new depth orders.

Currently, the depth order is written back to each actor's mSortedDepth,
and the nodes are updated with a single message containing the nodes
in depth order. (Could be used in a future patch to reduce renderer sorting)

Created a memory pool for managing the lifetime of the depth nodes that
exists for the lifetime of the application.

Change-Id: Iaed004eb8dc790fcc0f6f1b4822037054193b552