Ensured consistency of scene graph and render items
There was a timing issue where UpdateManager's updateScene could be
false, but the messageQueue flushed messages to the process queue just
prior to ProcessSMessages being called which updated the scene graph.

Have ensured that ProcessMessages returns true if there were updates
to the scene graph, and changed UpdateManager to process the scene
appropriately.

Change-Id: I773f5eae2820ad1d93a78d0f3c784a7592ea320a
Signed-off-by: David Steele <david.steele@samsung.com>