Fix the USE_ASSIGNED_SIZE resize policy to not overwrite the actor's original policy
OnRelayout is called during the relayout process, and if the actor has calculated
the size of child actors it will add them to the relayout container with their
desired size and set the ResizePolicy::USE_ASSIGNED_SIZE resize policy on them.
However, by doing this, it overwrites the original resize policy of the actor,
so the size calculation of the actor could be incorrect during any further
relayout process.

This patch fixes the above issue by setting an internal flag instead of overwriting
the original resize policy in order to notify the size negotiation algorithm to
assign the size of child actors as calculated during OnRelayout.

Change-Id: Ib1b2e4cfc2a39cfd653e64291e86508351d15606