Implemented basic CPU image masking
Given a PixelData containing an image and a PixelData to use as a mask,
this will apply the alpha channel of the mask to the image; either
multiplying into the image's alpha channel, or if it doesn't have one,
converting the image to RGBA888 and applying the mask's alpha channel.

If the mask is a different size than the image, then this will sample a
single pixel from a smaller mask, or use bilinear filtering to sample
from a larger mask.

Change-Id: I48b5d332627a4fb85899883868d278bcbc2b9c8e