Simplified actor depth traversal algorithms
Instead of having a complex sibling order and hierarchy depth combination,
changed to use actor order in mChildren; Sibling operations now move
the actor pointers in this array.

Automatically regenerates a depth index for every actor in the tree
when any part of the tree changes. (i.e. actor add/remove, raise/lower etc)
at the end of an event processing cycle, and resends a list of nodes
to UpdateManager in a single message.

Change-Id: I1f716132f0e3a13fc21fbebf188287dc893956ee
Signed-off-by: David Steele <david.steele@samsung.com>