Added mechanism for registering child properties on arbitrary actor
Want to register child properties for a container view from a second
class, e.g. a layout. This adds mechanism to register properties on
a registered type by type name.

Added a signal which triggers when a property is set on an object
(Note, OnPropertySet is available for derived types, but not for arbitrary objects)

Change-Id: Icf5c9856c66e1beadf80e55bc31ca9c1130f4a60
Signed-off-by: David Steele <david.steele@samsung.com>