[Vulkan] Image layout transition and texture test (WIP )
Added texture test that for now uploads raw 32bit data into an image
and prepares the image layout to be ready for sampling by the shader.

Note: it is still a test code which will be a base to build an actual
texture upload on the Graphics side ( for now just gathering all the
"building blocks" of the process ).