Fixing bug with changing a shader on a renderer
If the renderer's shader is updated, as is the case in the refraction demo,
then the texture doesn't get shown. This is because the attributes location
cache isn't updated on the renderer when the shader changes (The attributes
location cache should really be in the Program, not in the Renderer).

Modified  scene-graph-renderer to ensure that the geometry is resent to the
RenderRenderer - this sets a flag to re-load the attributes and cache them.

Change-Id: I8edfab92f89213ddb19ca3bd469e2be756c647ad
Signed-off-by: David Steele <david.steele@samsung.com>