Added ChildAdded and ChildRemoved signals to Actor
Actor currently has OnChildAdded and OnChildRemoved to inform derived classes when
children are added to or removed from the actor. However, there is currently no
mechanism for another object to find out this information. Added signals so that
other objects can listen for this state change.

Note, these signals are sent after the callbacks; if the derived classes choose
to do something odd as a result of the callback, such as reparent the child, the
signal may not reflect truth in the derived parent.

Change-Id: I348bf0a3314baaa6f7b1a1c939e3ce9821f3485a
Signed-off-by: David Steele <david.steele@samsung.com>