Fixed uniform mappings to handle actor properties
Uniform mappings weren't taking account of actor properties
(Removed in render-renderer.cpp deletion).

Updated scene-graph-renderer to ensure that all uniform maps are
properly collected into a single collectedUniformMap in the right
priority. This only happens when a property is registered on the
associated shader, renderer or node.

Only parses the current buffer of the collected uniform map now,
instead of both buffers.

Change-Id: I744f2c4494eb788b49117186d209e3b440db3fe9
Signed-off-by: David Steele <david.steele@samsung.com>