Fix to avoid render-task waiting forever
We require a finished signal to be emitted after rendering to the screen,
i.e. to know when the scene has been rendered for the 1st time.

Unfortunately RenderTask SetRefreshRate(REFRESH_ONCE) is causing scene updates to run forever.
This is is due to the mWaitingToRender flag which in that case will never be set to false.

Change-Id: Ia3cd411e65adc613be81db3d09d35ade87ae9f62