Changed command buffer inheritance
For the first command buffer in a render pass, the inherited render
pass info should come from the render pass, not the swapchain.

Modified Command buffer to always pass in inheritance info rather
than default it to nullptr - this forces users to think about what
inheritance info is needed, if any.

Change-Id: I99619094384e07f0943595ecd75a204cc11548a7