[vulkan] Add to refer window's size for vkGetPhysicalDeviceSurfaceCapabilities.
In Khronos spec for GetPhysicalDeviceSurfaceCapabilities,
This api does not have guarantee to get valid window size.
This patch is for invalid case.
The related spec's url and document are below.
"currentExtent is the current width and height of the surface,
or the special value (0xFFFFFFFF, 0xFFFFFFFF) indicating
that the surface size will be determined by the extent of a swapchain targeting the surface."

https://www.khronos.org/registry/vulkan/specs/1.0-extensions/html/vkspec.html#vkGetPhysicalDeviceSurfaceCapabilitiesKHR