[Bugfix] Pipeline cache wasn't deleting entries from mCacheMap
On pipeline remove, the actual pipeline was deleted but the hash map wasn't updated in case there were no more pipelines with specific hash value. As result, the mCacheMap was always growing and storing empty vector object for each hash value that didn't have a pipeline assigned. This now has been fixed and in case when the vector of pipelines is empty, the entry is erased from the cache map.

Change-Id: Ifb9bb42e0ee12ed212d6ed39723affcf9fdf31ec