[Vulkan] Wrapped every render pass in a separate command buffer
In order to preserve dependencies between render tasks, we need a synchronisation
mechanism between them to ensure that the first one finishes rendering before
the next one starts.

In theory, this should be achieveable with sub-render-pass dependencies, but
this was not the case.

Instead, we are going with a stronger synchronisation mechanism - in the first
case, we are wrapping each render pass within a separate command buffer, and
enforcing the timing using VkFence between each one.

This appears to work in the test application. (With the caveat that render-once
render tasks aren't currently programmed to wait for resources to be ready any
more - requires more work on the toolkit/application side).

Future work - move from VkFence to semaphores (Should be faster)

Change-Id: I2d4ac2d9f4f818830870c969304fe25ee02d0e21
Signed-off-by: David Steele <david.steele@samsung.com>