Fix to crash on exit.
On exit, the renderers are destroyed before render instructions. Since render instructions tries to destroy render commands attached to the renderers, it uses pointer that is already invalid and it results with crash. The FreeRenderCommands() function has been introduced in order to indicate whether we are shutting DALi down or not. Safet check is performed and the pointer to the renderer is set to nullptr if we performing the shutdown.

Change-Id: I98cbb25b539ecacf5bf2d6b0ba451dc1fb1b3e47