Fix the minimum pinch gesture distance
Previously, the minimum pinch distance is determined by dividing the screen height by 85,
which is proportional to the screen height.

The problem is that, on the big size screen (or law dpi) devices,
the user have to make a long distance between their fingers to make a pinch gesture,
it makes the user misunderstand that the pinch gesture is not workoring properly.

This patch fixes the problem by calculating minimum pinch distance using screen dpi.

Change-Id: Iaab2d37f43421d53f9aa14060bd55465df4f9343
Signed-off-by: Jiyun Yang <ji.yang@samsung.com>