Added small buffer optimization to Property::Value.
With SBO optimization we don't need to allocate memory
for the Pimpl pointer as we will be creating the object
inplace in th Storage buffer.
After this change the Property::Value size will be 16byte
instead of 8. I think the overhead is minimal comparing to
the number of allocation this class does in a dali application.

Note: with the SBO and union trick the implementaion is little
bit complex than before. but as this class is core for property
and animation and used every where. the complexity is justified
in my opinion.

Change-Id: Ibe14fe92f15c2ab93395c9731bd1fca902e98259