Const correctness improvements for Property::Value.
GetArray() and GetMap() return non-const pointers, which allow const
objects being modified. While DALi itself doesn't seem to abuse this
in client code it may still lead to hard to trace bugs or UB at worst.

It makes sense that the modification of an element of a Property::Map
or Array that a Property::Value holds, means modification of the
Property::Value itself.

- GetArray/Map() const has been changed to return pointer-to-consts;
- a non-const counterpart has been added to each;

Change-Id: I187770bbb3cf355fbf1df95c909bad3ddc8cdb23
Signed-off-by: György Straub <g.straub@partner.samsung.com>