Adding graphics sync object
Capture requires a pollable sync object so that we can
know when a texture has been written to on the GPU.

Added Graphics::SyncObject, which can be added to
CommandBuffer::END_RENDER_PASS. This is a pollable object
that will test positive when the render pass has been
flushed.

This is owned by RenderTracker; a container of which
is owned by RenderManager separately to the RenderInstruction
which communicates when a sync object is needed.

Renamed Integration::GlSyncAbstraction to Integration::GraphicsSyncAbstraction

Change-Id: I4b5aa27755f1261532a188a9235d088c8bd4ec57
Signed-off-by: David Steele <david.steele@samsung.com>