Add Timer at TapGestureRecognizer
Currently, when double tap operation is performed,
single tap and double tap are always transmitted together.

So, when single and double taps are required, then after one tap, we wait for the timeout,
if the timeout happens, then we emit the single-tap,
if we tap again during this time, then we emit a double-tap instead.

Change-Id: I4a5faecc8cf1cc50734b51ad1701e4823c44baab