Decrease property registration time
For animated custom properties attached to objects, such as uniforms
attached to visual renderers, the registration time is too long.

One of the culprits is the Object::RegisterProperty() method iterating
over all the preceding property names to ensure the uniqueness of the
property name.

In the visual case, we can remove this check, as we can ensure that
each call to RegisterProperty is only done once per uniform.

Added a method RegisterUniqueProperty (which should really be called
RegisterPropertyWithoutUniquenessCheck, but hey!) to perform the same
registration without the name lookup checks.

Change-Id: Id3b78e342415aef65f0ff121ae2c8ec4a9ec80a6
Signed-off-by: David Steele <david.steele@samsung.com>