Add a Hit-Test result events.
In the case of TouchEvent or Gesture, there is no way to propagate the event to the view below that is not related.

So, before sending an touch event, send an hitTestResult event to the view in the hit-test process to ask whether it will be hit or not.
If it returns false, it means that it will not be hit, and the hit-test continues to the next view.

Change-Id: I5b8681f569f5c3b89acc3fb23c1eada3d49b1c7b