Changed handling of uniforms of arrays of structs
A) For uniforms of the form "basename[index].element", the array index handling
didn't work.

In GL (on Ubuntu) the shader reflection for such uniforms provides
individual locations for each element of the array/struct;

B) For uniforms of the form "basename[index]", where basename is a basic
type (float/vecN, etc), the GL shader reflection produces only 1 location
for the basename, and provides the element count.

Mapping properties to such uniforms is done by having a Property per
element. For the first case, no index/array handling is needed.
For the second case, each property needs to match to the basename without
the array subscript, and also store it's array index.

Modified property setup to remove array index for properties of the first type.

Modified uniform lookup to handle the different hashes appropriately.

Modified test graphics to ensure the uniform reflection can be set up in the
same way as GL provides.

Change-Id: I4be92f3e6933ff1b9b4a7d48e97f5629930c0b4c
Signed-off-by: David Steele <david.steele@samsung.com>