Memory pool key uses ptr in 32bit mode
When compiled for a 32bit architecture, memory pool keys are
functionally identical to ptrs, and don't require any lookup or
comparison.

Have removed int32_t comparison operators and made the INVALID const
private. Added nullptr comparison operator to ensure that the keys can
be compared against nullptr, or utilize the boolean operator to check
for validity. (An int32_t key of 0 is valid in 64 bit mode)

Change-Id: I0eff1efede539e02c330ab601b10f018fcebeafd