Adding frame callback notification system
We need some way of synchronizing update properties set
from event side with properties set from frame callback.
For example, if frame callback is used to update physics
objects, and we want to sync with actor visibility, then
the actor visibility might occur after the first frame
callback adding a physics object.

This patch adds a synchronization point to the FrameCallback
UpdateProxy. The application can notify the FrameCallback that
it wants to set a sync point, and also modify dali object
properties. Then, in the update during which the associated
messages are being processed, that sync point is available
in the FrameCallback::Update() UpdateProxy. So, the application
can synchronize any other properties during the callback,
knowing they are happening in the same frame.

If multiple notifications are sent before the next Update,
all of them are queued with unique values.

(But, it's not a true sync point, it doesn't intersperse
any Update() callbacks during processing the message queue)

Change-Id: I154539044ceaff17c55c0640c3828010b087844a
Signed-off-by: David Steele <david.steele@samsung.com>