Change the TextureSet Count when we set empty texture or sampler
Let we change the number of textures / samplers in TextureSet
when we setting empty texture / sampler.

It will ignore some useless new creation of TextureSet if we just want to
remove some textures.

And also, let we ignore useless Sampler nullptr setup. We can control
cases when sampler is nullptr.

Change-Id: Iddc99bf2d7ad934a39e12ebe831ebffa8a7882b3
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>