Add API for check core thread + shutting down
Add some API to check whether the core is shutting down or not,
So out of dali-core side can also know it.

Also, Make combined helper API to check whether core installed before or not.
It will be useful when we check worker thread access into UI classes of dali.

Change-Id: I2d8d3120b4fc959056f4644846d3d14a3b89bcc6
Signed-off-by: Eunki Hong <eunkiki.hong@samsung.com>