[Tizen] Add FeedTouch api to GestureDetector.
This is a method of recognizing gestures by feeding a touch event to a specific actor.

This causes the actor to recognize the gesture without going through a hittest.

So, you can recognize gestures based on touchEvent as shown below.

mTapDetector = TapGestureDetector::New();
mTapDetector.DetectedSignal().Connect(this, &HelloWorldController::OnTap);
control.TouchedSignal().Connect(this, &HelloWorldController::OnTouch);

bool OnTouch(Actor actor, const TouchEvent& touch)
{
 mTapDetector.FeedTouch(actor, touch);
 return true;
}

This is the beginning of a new gesture recognition method.

Change-Id: Iecf0eab7a82e54a981f69d2e309675afcfa439f1