Keep node by id + Make SceneGraphTraveler don't travel anymore
Let we keep whole nodes cache by id.
And let we make scene graph traveler use this system instead of traveling.

Previously, if we want to check whether given node is under root node,
we should search whole childrens (travel up-to-down).

Instead, let we travel down-to-up, s.t. will spend much less time than before.

And also, let we make a way to user don't use root actor for frame callback.
It is not a neccessary if we ask to node to mUpdateManger.

Change-Id: Ied34d176a05c87687cbfe0530612d53606c7e7a4
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>