Remove inclusion of gl-abstraction.h
Remove all #includes.

Would like to move this abstraction to dali-adaptor, but many tests use the test-gl-abstraction
layer to check that e.g. uniforms have been written. Instead, they should be modified to use the
test graphics layer, but that's a lot of work :/.

So, have just removed the usage, but left the file there.

Change-Id: I16f4945eb54ede451f4c91ee130284e84d653309
Signed-off-by: David Steele <david.steele@samsung.com>