Do not travel scene tree what we don't need to
During Renderables adding cases, there was some hidden tree treversal exist
when the view visibility changed (false --> true).

This mUpdated = true; will be used when we check DamagedRects.

Since we dont need to propargate this informations to all decents,
let we propargate this visibility changed information only for this
recursive function, and forget.

Change-Id: I36e6e045a72d7bdfb1b0ea2d5b4784cc05e6173c
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>