Do not call RemoveObserver when we don't add it for PropertyResetter
It is possible that PropertyResetter destructed before Initialize()
when app terminate case.

If then, it is possible to got some crash when we try to access to
'destroyed' property owner, and call RemoveObserver();

To avoid this case, let we make ensure that we call initalized so
it is possible to call RemoveObserver().

Change-Id: I24b74f1d963090d867e1d047fe739132f9465ee3
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>