Make Transform Component dirty as 2 frames, not for true/false
Since ResetToBaseValue for TransfromProperty doesn't use basic
AnimatableProperty logic, the dirty flag is not match with common sence.

To match it as frame-by-frame, let we make component dirty as 2 bit,
and age-down every frame.

And after we allow to use dirty flag feature for transform,
we can determine whether we need to re-calculate matrix or not.

If that flag be used, we can 'skip' heavy matrix multiply operation.

Change-Id: Icf51e0be37962bee5147c08bb0cd3302bf203f89
Signed-off-by: Eunki, Hong <eunkiki.hong@samsung.com>